General QOL refactoring for repo
