+++ framework/set/provisioning/custom/rke1/setConfig.go
@@ -13,7 +13,6 @@ import (
 	"github.com/rancher/tfp-automation/framework/set/provisioning/custom/nullresource"
 	"github.com/rancher/tfp-automation/framework/set/resources/providers/aws"
 	"github.com/rancher/tfp-automation/framework/set/resources/providers/vsphere"
-	"github.com/sirupsen/logrus"
 )
 
 // SetCustomRKE1 is a function that will set the custom RKE1 cluster configurations in the main.tf file.
@@ -48,12 +47,6 @@ func SetCustomRKE1(terraformConfig *config.TerraformConfig, terratestConfig *con
 	SetRancher2Cluster(rootBody, terraformConfig, terratestConfig)
 
 	nullresource.CustomNullResource(rootBody, terraformConfig, terratestConfig)
-
-	_, err := file.Write(newFile.Bytes())
-	if err != nil {
-		logrus.Infof("Failed to write custom RKE1 configurations to main.tf file. Error: %v", err)
-		return nil, nil, err
-	}
 
 	return newFile, file, nil
 }
