+++ framework/set/provisioning/nodedriver/rke2k3s/setProxyConfig.go
@@ -7,19 +7,23 @@ import (
 	"github.com/zclconf/go-cty/cty"
 )
 
+const (
+	port = "3228"
+)
+
 // SetProxyConfig is a function that will set the proxy configurations in the main.tf file.
 func SetProxyConfig(clusterBlockBody *hclwrite.Body, terraformConfig *config.TerraformConfig) error {
 	agentVarsOneBlock := clusterBlockBody.AppendNewBlock(defaults.AgentEnvVars, nil)
 	agentVarsOneBlockBody := agentVarsOneBlock.Body()
 
 	agentVarsOneBlockBody.SetAttributeValue(defaults.ResourceName, cty.StringVal(httpProxy))
-	agentVarsOneBlockBody.SetAttributeValue(defaults.Value, cty.StringVal("http://"+terraformConfig.Proxy.ProxyBastion+":3228"))
+	agentVarsOneBlockBody.SetAttributeValue(defaults.Value, cty.StringVal("http://"+terraformConfig.Proxy.ProxyBastion+":"+port))
 
 	agentVarsTwoBlock := clusterBlockBody.AppendNewBlock(defaults.AgentEnvVars, nil)
 	agentVarsTwoBlockBody := agentVarsTwoBlock.Body()
 
 	agentVarsTwoBlockBody.SetAttributeValue(defaults.ResourceName, cty.StringVal(httpsProxy))
-	agentVarsTwoBlockBody.SetAttributeValue(defaults.Value, cty.StringVal("http://"+terraformConfig.Proxy.ProxyBastion+":3228"))
+	agentVarsTwoBlockBody.SetAttributeValue(defaults.Value, cty.StringVal("http://"+terraformConfig.Proxy.ProxyBastion+":"+port))
 
 	agentVarsThreeBlock := clusterBlockBody.AppendNewBlock(defaults.AgentEnvVars, nil)
 	agentVarsThreeBlockBody := agentVarsThreeBlock.Body()
